refactor: break into smaller components
